restructure handlers
